create categories-list.js
