+++ src/index.js
@@ -2,3 +2,4 @@ import './js/selectors.js';
 import './js/slider.js';
 import './js/modal.js';
 import './js/categories.js';
+import './js/categories-list.js';
